# crates/markitai-cli/src/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Markitai service API",
    "version": "set from the build when served",
    "summary": "Convert files and URLs to Markdown, manage saved jobs and model settings.",
    "description": "The REST API of `markitai serve`. Loopback clients are trusted; other clients send the startup token as `Authorization: Bearer` (preferred) or `?token=`. Every error body is `{detail, code, reason}`; present text by `reason`, then `code`. See docs/serve.md, docs/service-settings.md and docs/provider-management.md for the contracts this document summarizes.",
    "license": { "name": "MIT", "identifier": "MIT" }
  },
  "servers": [{ "url": "/", "description": "The service that serves this document" }],
  "security": [{}, { "bearerAuth": [] }, { "tokenQuery": [] }],
...
  "paths": {
...
    "/api/download-tickets": {
      "post": {
        "tags": ["service"],
        "operationId": "createDownloadTicket",
        "summary": "Issue a single-use, 60-second ticket for one download",
        "description": "For clients that must let the browser open a download natively (a navigation cannot send `Authorization`). The returned `url` carries `?ticket=` instead of the token; it admits exactly one GET of exactly `path` within `expires_in` seconds. A ticket presented for another path or method, or a second time, is refused with 401 `ticket_invalid`. At most 64 tickets are outstanding (429 `too_many_tickets`).",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TicketRequest" } } }
        },
        "responses": {
          "201": { "description": "Ticket issued", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Ticket" } } } },
          "401": { "$ref": "#/components/responses/Error" },
          "422": { "$ref": "#/components/responses/Error" },
          "429": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/jobs": {
      "post": {
        "tags": ["jobs"],
        "operationId": "createJob",
        "summary": "Start a job from uploaded files and/or URLs",
        "description": "Files precede URLs in the returned order. URL conversion needs a loopback peer or the token, also with --no-auth. A request that itself asks for model processing while no model is routable is refused with 422 `llm_unavailable` before anything is stored.",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "files": { "type": "array", "items": { "type": "string", "contentMediaType": "application/octet-stream" }, "description": "Repeated file fields; at most 100 MiB each" },
                  "urls": { "type": "string", "description": "A JSON array of http(s) URLs" },
                  "options": { "type": "string", "description": "A JSON object of JobOptions" }
                }
              },
              "encoding": { "files": { "contentType": "application/octet-stream" } }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "urls": { "type": "string", "description": "A JSON array of http(s) URLs" },
                  "options": { "type": "string", "description": "A JSON object of JobOptions" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Job created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreatedJob" } } } },
          "400": { "$ref": "#/components/responses/Error" },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" },
          "409": { "$ref": "#/components/responses/Error" },
          "413": { "$ref": "#/components/responses/Error" },
          "422": { "$ref": "#/components/responses/Error" },
          "503": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/jobs/{job_id}": {
      "get": {
        "tags": ["jobs"],
        "operationId": "getJob",
        "summary": "The job's current snapshot",
        "parameters": [{ "$ref": "#/components/parameters/JobId" }],
        "responses": {
          "200": { "description": "Snapshot", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/JobSnapshot" } } } },
          "401": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
    "/api/jobs/{job_id}/archive": {
      "get": {
        "tags": ["jobs"],
        "operationId": "downloadJobArchive",
        "summary": "Download every output file of a finished job as a ZIP",
        "security": [{}, { "bearerAuth": [] }, { "tokenQuery": [] }, { "ticketQuery": [] }],
        "parameters": [{ "$ref": "#/components/parameters/JobId" }],
        "responses": {
          "200": { "description": "ZIP archive", "content": { "application/zip": { "schema": { "type": "string", "contentMediaType": "application/zip" } } } },
          "401": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "409": { "$ref": "#/components/responses/Error" },
          "413": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/history": {
      "get": {
        "tags": ["history"],
        "operationId": "listHistory",
        "summary": "Saved jobs, newest first, including jobs other local processes added",
        "responses": {
          "200": { "description": "Summaries", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/HistoryEntry" } } } } },
          "401": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
    "/api/history/{job_id}": {
      "delete": {
        "tags": ["history"],
        "operationId": "deleteHistoryJob",
        "summary": "Delete a terminal saved job and its files",
        "parameters": [{ "$ref": "#/components/parameters/JobId" }],
        "responses": {
          "204": { "description": "Deleted" },
          "401": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "409": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/settings/llm": {
      "get": {
        "tags": ["settings"],
        "operationId": "getSettings",
        "summary": "Revision and secret-free saved and detected deployments",
        "responses": {
          "200": { "description": "Settings", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SettingsView" } } } },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "description": "The startup token (printed at start, or MARKITAI_SERVE_TOKEN). Loopback peers need none." },
      "tokenQuery": { "type": "apiKey", "in": "query", "name": "token", "description": "The startup token as a query parameter; accepted for compatibility, but it can end up in logs and history, so prefer the header." },
      "ticketQuery": { "type": "apiKey", "in": "query", "name": "ticket", "description": "A single-use download ticket from POST /api/download-tickets, valid for one GET of the path it was issued for within 60 seconds." }
    },
    "parameters": {
      "JobId": { "name": "job_id", "in": "path", "required": true, "schema": { "type": "string" } },
...
    },
    "responses": {
      "Error": { "description": "Refusal", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["detail", "code", "reason"],
        "properties": {
          "detail": { "description": "The service's English message, or an object for settings conflicts that carry `detail.code`", "oneOf": [{ "type": "string" }, { "type": "object" }] },
          "code": { "type": "string", "enum": ["bad_request", "unauthorized", "forbidden", "not_found", "method_not_allowed", "conflict", "payload_too_large", "invalid_request", "rate_limited", "unavailable", "server_error"] },
          "reason": {
            "type": "string",
            "description": "Stable machine code of the specific cause. The list is the set the service writes; new codes may be added.",
            "examples": ["ambiguous_legacy_model_name", "archive_too_large", "batch_pending", "config_changed", "config_missing", "config_too_large", "discovery_timeout", "empty_job", "enhancement_failed", "file_not_found", "file_too_large", "form_field_too_large", "history_empty", "host_not_allowed", "internal_error", "invalid_allowed_host", "invalid_json", "invalid_multipart", "invalid_options", "invalid_provider_request", "invalid_retry_body", "invalid_settings_request", "invalid_ticket_path", "invalid_url", "invalid_urls", "item_busy", "item_not_found", "job_id_collision", "job_not_found", "job_not_running", "job_running", "json_required", "llm_unavailable", "method_not_allowed", "no_output", "no_source", "not_retryable", "nothing_to_stop", "origin_not_allowed", "output_conflict", "output_identity_conflict", "persistence_failed", "provider_busy", "provider_unavailable", "remote_url_forbidden", "request_too_large", "result_too_large", "result_unavailable", "route_not_found", "settings_durability_unknown", "settings_entry_missing", "settings_forbidden", "settings_io_failed", "settings_read_only", "shutting_down", "stale_revision", "ticket_invalid", "token_required", "too_many_items", "too_many_tickets", "unknown_preset", "unsupported_url_scheme", "upload_missing"]
          }
        }
      },
...
      "TicketRequest": {
        "type": "object",
        "required": ["path"],
        "additionalProperties": false,
        "properties": { "path": { "type": "string", "description": "`/api/jobs/{job_id}/files/{relpath}`, `/api/jobs/{job_id}/archive` or `/api/history/archive`, exactly as it will be requested (percent-encoded ASCII, no query)", "examples": ["/api/jobs/0123456789ab/archive"] } }
      },
      "Ticket": {
        "type": "object",
        "required": ["ticket", "url", "expires_in"],
        "properties": {
          "ticket": { "type": "string", "description": "64 hexadecimal characters" },
          "url": { "type": "string", "description": "The path with `?ticket=`; open it once" },
          "expires_in": { "type": "integer", "description": "Seconds the ticket stays valid", "examples": [60] }
        }
      },
...
      "CreatedItem": {
        "type": "object",
        "required": ["item_id", "name", "kind"],
        "properties": { "item_id": { "type": "string" }, "name": { "type": "string" }, "kind": { "type": "string", "enum": ["file", "url"] } }
      },
      "CreatedJob": {
        "type": "object",
        "required": ["job_id", "items"],
        "properties": { "job_id": { "type": "string" }, "items": { "type": "array", "items": { "$ref": "#/components/schemas/CreatedItem" } } }
      },
...
      "JobProgress": {
        "type": "object",
        "required": ["status", "done", "failed", "total"],
        "properties": {
          "status": { "type": "string", "enum": ["running", "done", "error"] },
          "done": { "type": "integer" },
          "failed": { "type": "integer" },
          "total": { "type": "integer" },
          "persistence_error": { "type": "string" }
        }
      },
      "JobSnapshot": {
        "allOf": [
          { "$ref": "#/components/schemas/JobProgress" },
          {
            "type": "object",
            "required": ["job_id", "created_at", "finished_at", "options", "items"],
            "properties": {
              "job_id": { "type": "string" },
              "created_at": { "type": "string" },
              "finished_at": { "type": ["string", "null"] },
              "options": { "$ref": "#/components/schemas/JobOptions" },
              "items": { "type": "array", "items": { "$ref": "#/components/schemas/Item" } }
            }
          }
        ]
      },
...
      "HistoryEntry": {
        "type": "object",
        "required": ["job_id", "created_at", "status", "total", "done", "failed", "skipped", "names_preview", "kinds_preview", "size_bytes", "origin", "retryable"],
        "properties": {
          "job_id": { "type": "string" },
          "created_at": { "type": "string" },
          "finished_at": { "type": ["string", "null"] },
          "status": { "type": "string", "enum": ["running", "done", "error"] },
          "total": { "type": "integer" },
          "done": { "type": "integer" },
          "failed": { "type": "integer" },
          "skipped": { "type": "integer" },
          "llm_enhanced": { "type": "integer" },
          "cost_usd": { "type": ["number", "null"] },
          "pricing": { "$ref": "#/components/schemas/Pricing" },
          "names_preview": { "type": "array", "items": { "type": "string" } },
          "kinds_preview": { "type": "array", "items": { "type": "string", "enum": ["file", "url"] } },
          "duration_ms": { "type": ["integer", "null"] },
          "size_bytes": { "type": "integer" },
          "origin": { "type": "string", "enum": ["web", "cli"] },
          "retryable": { "type": "boolean" }
        }
      },
...
      "SettingsView": {
        "type": "object",
        "required": ["configured", "routable", "source", "config_path", "config_origin", "revision", "deployments", "detected"],
        "properties": {
          "configured": { "type": "boolean" },
          "routable": { "type": "boolean" },
          "source": { "type": "string", "enum": ["config", "detected", "none"] },
          "config_path": { "type": "string" },
          "config_origin": { "type": "string" },
          "revision": { "type": "string" },
          "deployments": { "type": "array", "items": { "$ref": "#/components/schemas/Deployment" } },
          "detected": { "type": "array", "items": { "$ref": "#/components/schemas/Deployment" } }
        }
      },
...
    }
  }
}